OBSDEF-14004: Restructured files, added common variables and styles
