+++ package.json
@@ -8,7 +8,7 @@
   "main": "dist/index.js",
   "types": "dist/index.d.ts",
   "dependencies": {
-    "@dellstorage/clarity-react": "^1.1.0",
+    "@dellstorage/clarity-react": "^1.1.8",
     "@types/node": "^12.14.1",
     "@types/react": "^17.0.0",
     "@types/react-dom": "^17.0.0",
@@ -127,7 +127,7 @@
     "jest-environment-jsdom-fourteen": "^0.1.0",
     "lint-staged": "^8.2.1",
     "mini-css-extract-plugin": "^1.0.0",
-    "node-sass": "^4.12.0",
+    "node-sass": "^6.0.1",
     "postcss": "^8.4.4",
     "postcss-loader": "^4.2.0",
     "prettier": "^2.5.1",
@@ -135,7 +135,7 @@
     "react-docgen-typescript-webpack-plugin": "^1.1.0",
     "react-scripts": "^4.0.3",
     "react-refresh": "^0.9.0",
-    "sass-loader": "^10.0.0",
+    "sass-loader": "^10.2.0",
     "style-loader": "^2.0.0",
     "ts-jest": "^24.0.0",
     "ts-loader": "^8.0.0",
