--- src/styles/Button.scss
@@ -1,84 +1,0 @@
-/**
- * Copyright (c) 2021 Dell Inc., or its subsidiaries. All Rights Reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- */
-
-@import "./Colors";
-@import "./Font";
-@import './button_mixin';
-
-.btn {
-        font-weight: 500;
-        width: max-content ;
-        height: 24px ;
-        border-radius: 2.5px;
-        line-height: 1.1rem;
-        text-transform: none;
-        letter-spacing: 0 !important;
-        border-color: $BLUE;
-        size: 16px;
-        font-size: 13px;
-   
-    btn-link{
-        @include link_button;
-    }
-
-    btn-icon {
-        min-width: 0;
-        color: $BLUE;
-        @include link_button;
-    } 
-
-    &:focus{
-        outline: none !important;
-        border-color: $LIGHTER-BLUE;
-    }
-
-    &:hover{
-        btn-primary{
-            background-color: $LIGHT-BLUE;
-        }       
-        border-color: $LIGHT-BLUE;
-    }
-
-    &:disabled{
-        border: none;
-        background-color: $WHITE;
-        color: $LIGHT-GREY;
-    }
-
-}
-
-    .btn.btn-primary{
-        min-width: 78px;
-        background-color: $BLUE !important;
-    }
-
-.warning-button{
-    background-color: $DARK-YELLOW !important;   
-    border-color: $DARK-YELLOW !important;
-    color: $BLACK !important;
-}
-
-.normal-button-size{
-    width: 82px;
-    height: 40px;
-    background-color: $BLUE;
-}
-
-.medium-button-size{
-    width: 82px;
-    height: 32px;
-    background-color: $BLUE;
-}
-
-.fluid-button-size{
-    width: 319px;
-    height: 40px;
-    background-color: $BLUE;
-}
