--- src/styles/Colors.scss
@@ -1,45 +1,0 @@
-/**
- * Copyright (c) 2021 Dell Inc., or its subsidiaries. All Rights Reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- */
-
-//Dell Design Standards Colors
-
-//Datagrid
-$datagrid-hover:#F7F7F7;
-$datagrid-selected:#E7F2FB;
-$datagrid-selected-hover:#D2E7F8;
-
-//Primary Colors
-$dell-gray: #E4E4E4;
-$dell-white: #FFFFFF;
-$dell-black: #333333;
-
-//Primary button colour
-$BLUE: #0076CE;
-$LIGHT-BLUE: #005DA4;
-$LIGHTER-BLUE: #6BACDE;
-$WHITE: #EEEEEE;
-
-
-//Warning button colour and text colour
-$DARK-YELLOW:  #F2AF00; 
-$BLACK: #333333; 
-
-//Input box colour
-$DARK-GREY: #888888;
-$LIGHT-GREY:  #BBBBBB;
-
-//Long Tag color
-$LIGHT-GRAY:#F5F5F5;
-
-//Error color
-$ERROR: #CE1126;
-
-//Helper text color
-$DARK-GREY-HELPER-TEXT: #666666;
