--- src/styles/_button_mixin.scss
@@ -1,23 +1,0 @@
-/**
- * Copyright (c) 2021 Dell Inc., or its subsidiaries. All Rights Reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- */
-
-@import './Colors';
-
-@mixin link_button {
-    margin-right: 0;
-    width: 48px ;
-    height: 40px ;
-    border-color: $BLUE;
-    background-color: transparent;
-    color: $BLUE;  
-    outline-color: $BLUE !important;
-    border: 1px ;
-    outline-style: solid; 
-}
