+++ src/styles/common/_borders.scss
@@ -1,0 +1,22 @@
+
+/**
+ * Copyright (c) 2022 Dell Inc., or its subsidiaries. All Rights Reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ */
+
+// Borders
+$border-default: 1px;
+$border-medium: 2px;
+$border-thick: 4px;
+
+// Button
+$btn-border-radius: 2.5px;
+
+//Input
+$input-focus-border-radius: 0rem;
+$input-disabled-border-radius: 0rem;
