+++ src/styles/common/_colors.scss
@@ -1,0 +1,69 @@
+/**
+ * Copyright (c) 2021 Dell Inc., or its subsidiaries. All Rights Reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ */
+
+// Dell Design Standard Color codes
+
+$primary-white: #FFFFFF;
+$gray-100: #F7F7F7;
+$gray-200: #EEEEEE;
+$gray-300: #E4E4E4;
+$gray-450:  #BBBBBB;
+$gray-550: #888888;
+$gray-700: #666666;
+$gray-900: #333333;
+
+$blue-50: #E7F2FB;
+$blue-450: #6BACDE;
+$blue-600: #0076CE;
+$blue-700: #005DA4;
+
+$yellow-300:  #F2AF00;
+
+//Error color
+$red-700: #CE1126;
+
+//Helper text color
+$subtext-color: $gray-700;
+
+//Button Colors
+$btn-default: $blue-600;
+$btn-hover: $blue-700;
+$btn-disabled: $gray-450;
+$btn-warning: $yellow-300;
+$btn-warning-font: $gray-900;
+$btn-link-border: $blue-600;
+
+// Disabled color
+$disabled-color: $gray-450;
+$disabled-background-color: $gray-200;
+
+//Foucs color
+$focus-border-color: $blue-450;
+
+// Outline
+$outline: $blue-600;
+
+// Grid colors
+$grid-text-default-color: $gray-900;
+$grid-border-color: $gray-300;
+$grid-backgroud-color: $primary-white;
+$grid-selected-background-color: $gray-300;
+$grid-hover-color: $gray-100;
+$grid-selected-color: $blue-50;
+
+// Checkbox color
+$checkbox-default-color: $blue-600;
+
+// Input colors
+$input-color: $gray-550;
+$input-placeholder-color: $gray-450;
+$input-border-color: $gray-550;
+$input-error-color: $red-700;
+$input-label-color: $gray-900;
