+++ src/styles/common/_fonts.scss
@@ -1,0 +1,35 @@
+/**
+ * Copyright (c) 2022 Dell Inc., or its subsidiaries. All Rights Reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ */
+
+//Primary Font
+$dell-font: -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, Oxygen, Ubuntu, Cantarell, 'Open Sans', 'Helvetica Neue', sans-serif;
+
+// Dell Design Standards Font Weights
+$font-weight-light: 300;
+$font-weight-regular: 400;
+$font-weight-medium: 500;
+
+// Font sizes
+$font-size-6: 0.6rem;
+$font-size-7: 0.7rem;
+
+// Grid
+$grid-font-size-default: 0.875rem;
+
+// Button
+$btn-font-size: 0.813rem;
+
+// Label
+$label-font-size: 0.875rem;
+
+// Input
+$subtext-font-size: 0.6rem;
+$input-label-font-size: 0.7rem;
+$input-placeholder-font-size: 0.7rem;
