+++ src/styles/common/_mixins.scss
@@ -1,0 +1,49 @@
+/**
+ * Copyright (c) 2022 Dell Inc., or its subsidiaries. All Rights Reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ */
+
+ @import "./../common/colors";
+ @import "./../common/layout";
+ @import "./../common/borders";
+ @import "./../common/fonts";
+
+
+ @mixin link_button {
+     margin-right: 0;
+     width: 48px ;
+     height: 40px ;
+     border-color: $btn-link-border;
+     background-color: transparent;
+     color: $btn-link-border;  
+     outline-color: $outline !important;
+     border: $border-default;
+     outline-style: solid; 
+ }
+ 
+ @mixin grid-title {
+    font-weight: $font-weight-medium;
+    color: $grid-text-default-color;
+    line-height: $grid-line-height;
+    font-size: $grid-font-size-default;
+}
+
+@mixin grid-cell {
+    font-weight: $font-weight-regular;
+    color: $grid-text-default-color;
+    font-size: $grid-font-size-default;
+    height: 46px;
+    padding-right: $grid-cell-padding-right;
+    padding-left: $grid-cell-padding-left;
+}
+
+@mixin btn-style($width, $height) {
+    width: $width;
+    height: $height;
+    background-color: $btn-default;
+}
